api/requirements.txt
fastapi
uvicorn
gunicorn
psycopg2-binary
slowapi
matplotlib
pandas